extract enums module
