+++ src/dsl/schema/deserialization.rs
@@ -147,7 +147,6 @@ where
 
 #[cfg(test)]
 mod tests {
-
     use super::*;
 
     #[test]
